pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "chord_metadata_service"  # can be renamed to katsu if inner module directory is renamed too
version = "4.0.0"
description = "An implementation of a clin/pheno metadata store for the Bento platform."
authors = [
    "Ksenia Zaytseva",
    "David Lougheed <david.lougheed@mail.mcgill.ca>",
    "Simon Chénard <simon.chenard2@mcgill.ca>",
    "Romaine Grégoire",
    "Paul Pillot <paul.pillot@computationalgenomics.ca>",
    "Son Chau",
]
readme = "README.md"
packages = [{include = "chord_metadata_service"}]
include = [
    "chord_metadata_service/chord/workflows/wdls/*.wdl",
    "chord_metadata_service/chord/tests/*.json",
    "chord_metadata_service/dats/*",
    "chord_metadata_service/mcode/tests/*.json",
    "chord_metadata_service/restapi/tests/*.json",
]
repository = "https://github.com/bento-platform/katsu"
license = "LGPL-3.0-only"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = "^3.8.1"
bento-lib = {extras = ["django"], version = "^6.0.1"}
Django = "^4.2.1"
django-autocomplete-light = "^3.9.4"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
djangorestframework = "^3.14.0"
djangorestframework-camel-case = "^1.3.0"
drf-spectacular = "^0.25.1"
elasticsearch = "~7.8.0"
fhirclient = "~3.2"
jsonschema = {version = "~4.17.3", extras = ["format"]}
psycopg2-binary = "~2.9.5"
PyJWT = {extras = ["crypto"], version = "^2.6.0"}
python-dateutil = "^2.8.2"
python-dotenv = "^0.21.1"
rdflib = "^6.2.0"
requests = "^2.31.0"
rfc3987 = "^1.3.8"
strict-rfc3339 = "^0.7"
tabulate = "^0.9.0"
uritemplate = "^4.1.1"
pyhumps = "^3.8.0"
adrf = "^0.1.1"

[tool.poetry.group.dev.dependencies]
coverage = "^6.5.0"
flake8 = "^6.0.0"
tox = "^4.2.8"
debugpy = "^1.6.4"
